Move the renderer's LivenessScope into the context
