+++ plugins/ui/src/deephaven/ui/_internal/__init__.py
@@ -3,8 +3,11 @@ from .RenderContext import (
     StateKey,
     StateUpdateCallable,
     OnChangeCallable,
+    InitializerFunction,
+    UpdaterFunction,
+    get_context,
+    NoContextException
 )
-from .shared import get_context, set_context, NoContextException
 from .utils import (
     get_component_name,
     get_component_qualname,
