+++ plugins/ui/src/deephaven/ui/hooks/use_render_queue.py
@@ -1,5 +1,4 @@
-from .._internal import OnChangeCallable
-from .._internal.shared import get_context
+from .._internal import OnChangeCallable, get_context
 
 
 def use_render_queue() -> OnChangeCallable:
